+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Autocompletions on invocations for file enums (For instance, now an Item File's Id will complete as a list of material also when pression ctrl + space)
+- DamageCause completion for DamageModifiers
 
 ### Fixed
 - Curly brackets autoclosing too late inside attribute values